PN-8910 esternalizzata gestione PNAG012
